requirements.txt
pygetwindow==0.0.9
pyautogui==0.9.54
opencv-python==4.8.0.76
numpy==1.26.0
pytesseract==0.3.10
pillow==9.5.0
keyboard
Pillow
